Optimize UI
Make the windows shown at the center of the screen
Add titles of the windows
Change some fonts

+++ Restaurant/src/net/fooddelivery/restaurant/win/UpdateFoodWin.java
@@ -12,7 +12,9 @@ import javax.swing.JTextField;
 import javax.swing.filechooser.FileNameExtensionFilter;
 
 import java.awt.Font;
+import java.awt.GraphicsEnvironment;
 import java.awt.Image;
+import java.awt.Point;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
@@ -22,7 +24,7 @@ import net.fooddelivery.restaurant.func.*;
 import net.fooddelivery.restaurant.models.Food;
 public class UpdateFoodWin {
 
-	private JFrame frame;
+	private JFrame frmUpdateFood;
 	private JTextField name,des,FileAdd;
 	private byte[]bFile;
 	private JTextField price;
@@ -37,7 +39,7 @@ public class UpdateFoodWin {
 			public void run() {
 				try {
 					UpdateFoodWin window = new UpdateFoodWin(foodman);
-					window.frame.setVisible(true);
+					window.frmUpdateFood.setVisible(true);
 				} catch (Exception e) {
 					e.printStackTrace();
 				}
@@ -57,37 +59,41 @@ public class UpdateFoodWin {
 	 * Initialize the contents of the frame.
 	 */
 	private void initialize() {
-		frame = new JFrame();
-		frame.getContentPane().setFont(new Font("SimSun-ExtB", Font.PLAIN, 18));
-		frame.setBounds(100, 100, 800, 600);
-		frame.setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
-		frame.getContentPane().setLayout(null);
+		frmUpdateFood = new JFrame();
+		frmUpdateFood.setTitle("Update Food");
+		final int width=800;
+		final int height=600;
+		Point p = GraphicsEnvironment.getLocalGraphicsEnvironment().getCenterPoint();
+	    frmUpdateFood.setBounds(p.x - width / 2, p.y - height / 2, width, height); 
+		frmUpdateFood.getContentPane().setFont(new Font("SimSun-ExtB", Font.PLAIN, 18));
+		frmUpdateFood.setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
+		frmUpdateFood.getContentPane().setLayout(null);
 		
 		JLabel lblNewLabel = new JLabel("Name: ");
 		lblNewLabel.setFont(new Font("SimSun-ExtB", Font.PLAIN, 18));
 		lblNewLabel.setBounds(81, 75, 72, 18);
-		frame.getContentPane().add(lblNewLabel);
+		frmUpdateFood.getContentPane().add(lblNewLabel);
 		
 		name = new JTextField();
 		name.setBounds(217, 74, 232, 24);
-		frame.getContentPane().add(name);
+		frmUpdateFood.getContentPane().add(name);
 		name.setColumns(10);
 		name.setText(foodman.food.getName());
 		
 		JLabel lblPhoto = new JLabel("Photo:");
 		lblPhoto.setFont(new Font("SimSun-ExtB", Font.PLAIN, 18));
 		lblPhoto.setBounds(81, 129, 72, 18);
-		frame.getContentPane().add(lblPhoto);
+		frmUpdateFood.getContentPane().add(lblPhoto);
 		
 		FileAdd = new JTextField();
 		FileAdd.setText(" Choose a image file from disk");
 		FileAdd.setBounds(217, 228, 328, 24);
-		frame.getContentPane().add(FileAdd);
+		frmUpdateFood.getContentPane().add(FileAdd);
 		FileAdd.setColumns(10);
 		
 		JLabel Photo = new JLabel("New label");
 		Photo.setBounds(285, 115, 100, 100);
-		frame.getContentPane().add(Photo);
+		frmUpdateFood.getContentPane().add(Photo);
 		ImageIcon temp=new ImageIcon(foodman.food.getPhoto());
 		ImageIcon icon = new ImageIcon(temp.getImage().getScaledInstance(Photo.getWidth(), Photo.getHeight(), Image.SCALE_DEFAULT));
 		bFile=foodman.food.getPhoto();
@@ -101,7 +107,7 @@ public class UpdateFoodWin {
 			    FileNameExtensionFilter filter = new FileNameExtensionFilter(
 			        "JPEG", "jpg");
 			    chooser.setFileFilter(filter);
-			    int returnVal = chooser.showOpenDialog(frame);
+			    int returnVal = chooser.showOpenDialog(frmUpdateFood);
 			    if(returnVal == JFileChooser.APPROVE_OPTION) {
 			    	FileAdd.setText(chooser.getSelectedFile().getPath());
 
@@ -122,39 +128,39 @@ public class UpdateFoodWin {
 			}
 		});
 		btnChooseFile.setBounds(569, 227, 145, 27);
-		frame.getContentPane().add(btnChooseFile);
+		frmUpdateFood.getContentPane().add(btnChooseFile);
 		
 		JLabel lblDescription = new JLabel("Description:");
 		lblDescription.setFont(new Font("SimSun-ExtB", Font.PLAIN, 18));
 		lblDescription.setBounds(81, 291, 108, 18);
-		frame.getContentPane().add(lblDescription);
+		frmUpdateFood.getContentPane().add(lblDescription);
 		
 		des = new JTextField();
 		des.setBounds(217, 290, 485, 24);
-		frame.getContentPane().add(des);
+		frmUpdateFood.getContentPane().add(des);
 		des.setColumns(10);
 		des.setText(foodman.food.getDescription());
 		
 		JLabel lblPrice = new JLabel("Price:");
 		lblPrice.setFont(new Font("SimSun-ExtB", Font.PLAIN, 18));
 		lblPrice.setBounds(81, 384, 108, 18);
-		frame.getContentPane().add(lblPrice);
+		frmUpdateFood.getContentPane().add(lblPrice);
 		
 		price = new JTextField();
 		price.setBounds(161, 383, 100, 24);
-		frame.getContentPane().add(price);
+		frmUpdateFood.getContentPane().add(price);
 		price.setColumns(10);
 		price.setText(new Double(foodman.food.getPrice()).toString());
 		
 		JLabel lblCalorie = new JLabel("Calorie:");
 		lblCalorie.setFont(new Font("SimSun-ExtB", Font.PLAIN, 18));
 		lblCalorie.setBounds(351, 386, 108, 18);
-		frame.getContentPane().add(lblCalorie);
+		frmUpdateFood.getContentPane().add(lblCalorie);
 		
 		calorie = new JTextField();
 		calorie.setColumns(10);
 		calorie.setBounds(473, 383, 100, 24);
-		frame.getContentPane().add(calorie);
+		frmUpdateFood.getContentPane().add(calorie);
 		calorie.setText(new Integer(foodman.food.getCalorie()).toString());
 		
 		JButton btnUpdate = new JButton("Update");
@@ -175,22 +181,31 @@ public class UpdateFoodWin {
 					foodman.UpdateFood(newf);
 					lock();
 					JOptionPane.showMessageDialog(null, "Information of Food Updated Successfully!", "Success!", JOptionPane.INFORMATION_MESSAGE);
-					frame.dispose();
+					frmUpdateFood.dispose();
 				}
 			}
 		});
 		btnUpdate.setBounds(415, 495, 113, 27);
-		frame.getContentPane().add(btnUpdate);
+		frmUpdateFood.getContentPane().add(btnUpdate);
 		
 		JButton btnDelete = new JButton("Delete");
 		btnDelete.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				foodman.DeleteFood();
-				frame.dispose();
+		        int n = JOptionPane.showConfirmDialog(null,"Are you Sure to Delete this Food?","Confirm Delete", JOptionPane.YES_NO_OPTION);
+		        if(n==JOptionPane.YES_OPTION){
+					foodman.DeleteFood();
+					frmUpdateFood.dispose();
+					JOptionPane.showMessageDialog(null, "The Food has been Deleted!", "Success!", JOptionPane.INFORMATION_MESSAGE);
+
+		        }
+		        else if(n==JOptionPane.NO_OPTION){
+		        	
+		        }
+
 			}
 		});
 		btnDelete.setBounds(569, 495, 113, 27);
-		frame.getContentPane().add(btnDelete);
+		frmUpdateFood.getContentPane().add(btnDelete);
 		lock();
 	}
 	private void lock(){
